requirements.txt
requests
waveshare-epd
python-dotenv
pillow